ml_service/requirements.txt
fastapi
uvicorn[standard]
requests
opencv-python-headless
mediapipe
numpy
python-dotenv
scikit-learn